06_cloud_security/StartupExcuses_monitoring/StartupExcuses/requirements.txt
Flask
psycopg2-binary
python-dotenv
dotenv
boto3
